requirements.txt
flask
flask_cors
pickle5
transformers
torch
Pillow
sentence-transformers